+++ examples/custom-autoloader.php
@@ -33,4 +33,12 @@ $script = $renderer($finder, $generator);
 echo $script;
 
 // default format is PNG
-echo $generator->createImageFile($renderer->getGraph()) . ' file generated' . PHP_EOL;  // @phpstan-ignore-line
+if (isset($argv[1])) {
+    // target folder provided
+    $cmdFormat = '%E -T%F %t -o ' . rtrim($argv[1], DIRECTORY_SEPARATOR) . '/custom_autoloader.graphviz.%F';
+} else {
+    $cmdFormat = '';
+}
+$graph = $renderer->getGraph();
+$target = $generator->createImageFile($graph, $cmdFormat);
+echo (empty($target) ? 'no' : $target) . ' file generated' . PHP_EOL;